// src/types/enums.d.ts
export enum ItemTypes {
    Sticker = 'sticker',
    Background = 'background',
    Widget = 'widget',
    Note = 'note'
}

export enum Categories {
    General = 'general'
}

export enum AssetExtensions {
    Gif = 'gif',
    Png = 'png',
    Jpg = ''
}

export enum Skins {
    Default = 'default',
    Golden = 'golden',
    Metal = 'metal',
    Notepad = 'notepad',
    Speech_Bubble = 'speech bubble',
    Stickie_Note = 'stickie note',
    HC_Bling = 'HC Bling',
    HC_Scifi = 'HC Scifi'
}

export enum ActionTypes {
    Edit = 'edit',
    Save = 'save',
    Cancel = 'cancel',
    Add = 'add',
    Generic = 'generic'
}